+++ Tests/Vendors/UIKit/UICollectionViewCell/ReusableClassCollectionViewCell.swift
@@ -1,0 +1,14 @@
+//
+//  ReusableClassCollectionViewCell.swift
+//  SwiftyUtils
+//
+//  Created by Tom Baranes on 01/07/2020.
+//  Copyright © 2020 Tom Baranes. All rights reserved.
+//
+
+import UIKit
+import SwiftyUtils
+
+final class ReusableClassCollectionViewCell: UICollectionViewCell, Reusable {
+
+}