+++ Tests/Vendors/UIKit/UICollectionViewCell/ReusableNibCollectionReusableView.swift
@@ -1,0 +1,15 @@
+//
+//  ReusableNibCollectionViewCell.swift
+//  SwiftyUtils
+//
+//  Created by Tom Baranes on 01/07/2020.
+//  Copyright © 2020 Tom Baranes. All rights reserved.
+//
+
+import UIKit
+import SwiftyUtils
+
+final class ReusableNibCollectionReusableView: UICollectionReusableView, NibReusable {
+
+    @IBOutlet private(set) weak var label: UILabel!
+}